Change #initialize in WarGame to accept Parameters
Because of the Server/Socket additions that will be made, it appears
as though the server will need to create a new player to associate
with each client. Therefore, it will need to create a game with those
players already made. The parameters in the WarGame class have new
players as default values.