Fix Tie Error
There was a possibility that, after coming to a tie, one of the
players would not have another card to play. I edited the code to
check for the game winner every time the play_round method was called.
The check for game winner method would check if any player did not have
any cards left in their hand, the other player wins, and implicitly
returns false. play_round only runs the standard game loop if that
check method evaluates to true.