Create Tests and add function to start game method
The start game method did not have any functionality to check if there
where enough players to start a game or give feedback to indicate so.
I wrote/rewrote and sorted the test suite for clarity. I also added
the needed fucntionality to make the start game method work correctly.

As a side, I aslo added the new runner script to the lib folder and
began writing a test for the play round method.