Add Initial Run_game and ask_ready
I created the run_game method that only calls ask_ready and a test to do
so. I then created a ask ready test that asks all the players in the
game if they are ready. I had to make an array in the WarGame that
stored all of the players. I also wrote a test to make sure it puts
the message to the players given.